requirement.txt
django
django-rest-framework
psycopg2-binary
pillow
